requirements.txt
Django==6.0.3
djangorestframework==3.17.1
django-cors-headers==4.9.0
gunicorn==25.3.0
whitenoise==6.7.0
requests==2.32.3
python-dotenv==1.2.2
